backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
asn1crypto==1.5.1
boto3==1.42.7
botocore==1.42.7
brotli==1.2.0
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
distro==1.9.0
dj-database-url==3.0.1
Django==6.0
django-cors-headers==4.9.0
django-csp==4.0
django-db-connection-pool==1.2.6
django-extensions==4.1
django-auditlog==3.0.0
django-storages==1.14.6
django-axes==6.3.0
django-permissions-policy==4.20.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
Faker==38.2.0
fonttools==4.61.0
greenlet==3.3.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
jiter==0.12.0
jmespath==1.0.1
MarkupSafe==3.0.3
numpy==2.3.5
oauthlib==3.3.1
openai==2.9.0
packaging==25.0
pgvector==0.4.2
pillow==12.0.0
polars==1.36.1
polars-runtime-32==1.36.1
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydyf==0.12.1
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-http-client==3.3.7
pytz==2025.2
pyotp==2.9.0
requests==2.32.5
requests-oauthlib==2.0.0
reportlab==4.3.1
s3transfer==0.16.0
sendgrid==6.12.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlparams==6.2.0
sqlparse==0.5.4
starkbank-ecdsa==2.2.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
weasyprint==67.0
webauthn==2.7.0
webencodings==0.5.1
Werkzeug==3.1.4
whitenoise==6.11.0
xlsxwriter==3.2.9
zopfli==0.4.0
